+++ src/Silverback.Integration/Messaging/Encryption/DecryptorConsumerBehavior.cs
@@ -47,7 +47,7 @@ public class DecryptorConsumerBehavior : IConsumerBehavior
 
         string? keyIdentifier = null;
 
-        if (envelope.Endpoint.Configuration.Encryption is SymmetricDecryptionSettings { KeyProvider: { } })
+        if (envelope.Endpoint.Configuration.Encryption is SymmetricDecryptionSettings { KeyProvider: not null })
             keyIdentifier = envelope.Headers.GetValue(DefaultMessageHeaders.EncryptionKeyId);
 
         envelope.RawMessage = _streamFactory.GetDecryptStream(envelope.RawMessage, envelope.Endpoint.Configuration.Encryption, keyIdentifier);