+++ docs/releases.md
@@ -73,6 +73,7 @@ uid: releases
     * Renamed Inbound namespace to Consuming and Outbound to Producing
     * Split IMessageSerializer / IMessageDeserializer and rename UseFixedType to UseModel
     * Removed Ping health check
+    * Kafka key not set to random Guid anymore (will be null if not explicitly set)
 * `BatchSettings`, `ChunkSettings`, `SequenceSettings`, ` EncryptionSettings`, etc. renamed to `BatchConfiguration`, `ChunkConfiguration`, `SequenceConfiguration`, ` EncryptionConfiguration`, etc.
 * Deprecated EF package
 * Changed integration of domain entities / domain events with DbContext